Refactor: extract getSingleUser and getTotalTweets function on helpers
